Create if-else that compares player and computer
Created a conditional that compares all combinations of playerSelection
to computerSelection. This way, each outcome can have specific comments.

Also included an else that will happen if the playerSelection and
computerSelection do not match up with any of the combinations. In that
case, the user inputted an invalid string and it will display a
notification.