Fix syntax and logic errors within program
First, I fixed the order of the conditional that determines who gets a
point. This way, if either won before doing nothing and moving to the
next iteration. This avoids not giving winner points.

Next, I declared the variables within the for loop. This way, it would
change the computerSelection and give the user another chance each
round. Before, it would just take the initial values and repeat five
times.

Finally, I declared roundWinner within the for loop, because without a
variable, the return value of playRound() would be lost.