Add visual confirmation of choices and game text
Added divs in the index.html file that display the game information in
real time. This increases interactibility because it shows the user what
they chose as well as the computer.

This removes the mysterious aspect of the computer and shows the user
why and how they lost, which is important in a game.